Updates: Compilation passed but seg fault during runtime
