+++ kokkosConfig.h
@@ -13,7 +13,7 @@ using KokkosMemSpace = Kokkos::HostSpace;
 #define KOKKOS_NAMESPACE kokkos_openmp
 #elif defined DIGI_KOKKOS_CUDA
 using KokkosExecSpace = Kokkos::Cuda;
-using KokkosMemSpace = Kokkos::CudaUVMSpace;
+using KokkosMemSpace = Kokkos::CudaSpace;
 #define KOKKOS_NAMESPACE kokkos_cuda
 #endif
